backend/requirements.txt
# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# AI / LLM
langgraph>=0.1.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
anthropic>=0.25.0

# Database
psycopg2-binary>=2.9.9      # PostgreSQL driver (Docker & local)

# HTTP Client
httpx>=0.27.0

# Environment
python-dotenv>=1.0.0

# Utilities
pydantic>=2.6.0
pydantic-settings>=2.2.0
